requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.10.0
supabase==2.3.0
alpaca-py==0.18.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
apscheduler==3.10.4
httpx==0.24.1
pydantic==2.5.3
pydantic-settings==2.1.0
feedparser==6.0.11
